python/requirements.txt
# Cortex Capital — AI-native hedge fund prototype stack
openbb[all]>=4.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
ib_insync>=0.9.86
groq>=0.11.0
anthropic>=0.40.0
streamlit>=1.40.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
yfinance>=0.2.40
plotly>=5.18.0
pydantic>=2.0.0
typing_extensions>=4.12.0
# Optional: pip install sqlite-vec  (native extension; see https://github.com/asg017/sqlite-vec)
